Deterministic stubbing 🎉
Allow injection of a date scheduler which will, when stubbing, be used instead
of the default behavior for executing stub closures.

This commit required reworking the internals of MoyaProvider, mostly factoring
out some common code to be shared by the subclass into a separate few methods
that are final and internal, and documented.