[Documentation] Improve setup example
This improves the setup example by adding a more detailed and common
use case to the doc/Examples.md. The reason behind this is that previously
it was not intuitively clear how the default behavior of Moya was and how all
parts came together to result to a request. This tries to solve that and also
makes the example more copy & pastable (which is always good in docs).