Provide default empty implementations for PluginType
Sometimes a custom`PluginType` doesn't need to do anything in `willSendRequest` or `didReceiveResponse` but has to provide an empty implementation to conform. Adding a default implementation in an extension leaves the custom plugin free to implement either, none, or both functions.

Not quite sure if this is a good idea or not for the framework  but I've found it useful in my Moya-based projects.