In Demo, Change completion using GitHub.userRepositories to use dematerialize()
This creates a much cleaner completion block by using the Result method `dematerialize()` in the same do-try-catch block as the response mapping. Adds a `mapNSArray()` extension to Response in GitHubAPI.swift to support this and as an example (see #950). Also an example of how to create a error message from an error with default.