[CI] Add new script to exit on non-trivial changes (docs, markdown, etc)
Working towards #1286, created a new script which lists the file names of anything changed within the provided commit.

As per CircleCI's docs (https://circleci.com/docs/1.0/environment-variables/) I've gotten the "CIRCLE_SHA1" variable which should be the commit hash.

Following that I do an inverted grep which searches for an entry that doesn't match the regex (so source file changes etc), and in that occasion will exit with code 0 (success).

In the event it does find something, it'll return with exit code 1 which I'm hoping is enough for Circle to stop on.

For testing purposes I've added some prints to see what's happening, the regex isn't final as will need to find out what we want to search for and include/exclude!