Halt further steps if changes only include trivial changes
Command was found from this community post: https://discuss.circleci.com/t/ability-to-return-successfully-from-a-job-before-completing-all-the-next-steps/12969/6 - I've also added this shell script to the regex with the intention that CI would (in an ideal world) ignore this commit and not do a full test. Honestly, haven't got hopes for this as I think it'll just skip this set of tasks - may need to move this as a pre step on the test stage.