Use new Alamofire append(_ fileURL, …) function with optional parameters
Before we had to specify both a `fileName` and a `mimeType`. If one of them was missing the other would be discarded resulting in potential bugs. In our case we were removing invalid characters to create a valid server fileName, but was getting discarded by this method resulting on failing uploads.