In a project I'm working on, we had a problem to navigate between the pathes, because it is a back-end driven-ui solution and there we from the json the path we should navigate to.
So basically, this path tells the whole url itself, without separating the query params.
So we get into that problem of having %3f in the place of ?
Debugging, we noticed that changing

from             self = target.baseURL.appendingPathComponent(targetPath)
to             self = URL(string: target.baseURL.absoluteString + target.path)!

It would work very well.
So this is the solution I'm suggesting.